filiere

+++ src/main/java/projet/ais/repository/ActeurRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projet.ais.models.Acteur;
 
-public interface ActeurRepository extends JpaRepository<Acteur, Long>{
+public interface ActeurRepository extends JpaRepository<Acteur, Integer>{
 
      Acteur findByIdActeur(long idActeur);
